+++ src/index.html
@@ -30,7 +30,7 @@
     </div>
   </main>
 
-  <script type="module" src="./index.html"></script>
+  <!-- <script type="module" src="./script.js"></script> -->
 </body>
 
 </html>
