+++ src/scripts/render/attender-counter.js
@@ -1,0 +1,6 @@
+import { localArrOfPersons } from "../../app"
+
+export function renderAtendersCount() {
+  const attenderCountContainer = document.getElementById('attendees-count')
+  attenderCountContainer.textContent = `Attendees (${localArrOfPersons.length})`
+}
